requirements.txt
orjson
discord.py
motor
